Rename `FileUtil` to `FileService`
