+++ runner/src/main/java/dev/amrw/runner/chain/run/command/CopyJarFromBuildContainer.java
@@ -4,7 +4,7 @@ import dev.amrw.runner.chain.run.RunChainCommandBase;
 import dev.amrw.runner.chain.run.RunChainContext;
 import dev.amrw.runner.exception.ContainerArchiveCopyingException;
 import dev.amrw.runner.service.DockerClientService;
-import dev.amrw.runner.util.FileUtil;
+import dev.amrw.runner.service.FileService;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.chain.Command;
 import org.apache.commons.chain.Context;
@@ -17,16 +17,16 @@ import java.io.IOException;
 @Log4j2
 public class CopyJarFromBuildContainer extends RunChainCommandBase {
 
-    private final FileUtil fileUtil;
+    private final FileService fileService;
 
     public CopyJarFromBuildContainer() {
         super();
-        fileUtil = FileUtil.getInstance();
+        fileService = FileService.getInstance();
     }
 
-    public CopyJarFromBuildContainer(final DockerClientService dockerClientService, final FileUtil fileUtil) {
+    public CopyJarFromBuildContainer(final DockerClientService dockerClientService, final FileService fileService) {
         super(dockerClientService);
-        this.fileUtil = fileUtil;
+        this.fileService = fileService;
     }
 
     // TODO: Add logic to skip this step if the main container already exists and `--rebuild` is false
@@ -50,11 +50,11 @@ public class CopyJarFromBuildContainer extends RunChainCommandBase {
         ) {
             // Make sure the directory exists, in case it's the first run.
             log.debug("Creating directory (path={})", RunChainContext.HOST_BIN_PATH);
-            fileUtil.mkdir(RunChainContext.HOST_BIN_PATH);
+            fileService.mkdir(RunChainContext.HOST_BIN_PATH);
 
             log.debug("Saving InputStream to file (source = {}, destination={})",
                     RunChainContext.GRADLE_LIBS_PATH, RunChainContext.HOST_GRADLE_LIBS_ARCHIVE_PATH);
-            fileUtil.toFile(archiveStream, RunChainContext.HOST_GRADLE_LIBS_ARCHIVE_PATH);
+            fileService.toFile(archiveStream, RunChainContext.HOST_GRADLE_LIBS_ARCHIVE_PATH);
         } catch (final IOException exception) {
             throw new ContainerArchiveCopyingException(
                     RunChainContext.GRADLE_LIBS_PATH, RunChainContext.HOST_GRADLE_LIBS_ARCHIVE_PATH);
